+++ R/elco-doc.R
@@ -1,6 +1,0 @@
-#' Internal vctrs methods
-#'
-#' @import vctrs
-#' @keywords internal
-#' @name elco-vctrs
-NULL