+++ R/elco_irms_import_csv.R
@@ -36,20 +36,22 @@ elco_irms_import_csv <- function(files) {
   d <- purrr::map(files, utils::read.csv, header = TRUE, as.is = TRUE, row.names = NULL)
 
   # extract analysis metadata
-  d_metadata_analysis <- purrr::map2_df(d, seq_along(d), function(x, i) {
-    tibble::tibble(
-      file_id = i,
-      file_name = x[1, 2, drop = TRUE],
-      measurement_time = lubridate::parse_date_time(paste0(x[1, 4, drop = TRUE], " ", x[1, 6, drop = TRUE]), orders = "dmy %h:%M"),
-      instrument_name = x[1, 8, drop = TRUE],
-      measurement_software_version = x[1, 10, drop = TRUE],
-      reanalysis_time = lubridate::parse_date_time(paste0(x[1, 12, drop = TRUE], " ", x[1, 14, drop = TRUE]), orders = "dmy %h:%M"),
-      reanalysis_software_version = x[1, 10, drop = TRUE]
+  d_metadata_analysis <-
+    purrr::map2_df(d, seq_along(d), function(x, i) {
+      tibble::tibble(
+        file_id = i,
+        file_name = x[1, 2, drop = TRUE],
+        measurement_time = lubridate::parse_date_time(paste0(x[1, 4, drop = TRUE], " ", x[1, 6, drop = TRUE]), orders = "dmy %h:%M"),
+        instrument_name = x[1, 8, drop = TRUE],
+        measurement_software_version = x[1, 10, drop = TRUE],
+        reanalysis_time = lubridate::parse_date_time(paste0(x[1, 12, drop = TRUE], " ", x[1, 14, drop = TRUE]), orders = "dmy %h:%M"),
+        reanalysis_software_version = x[1, 10, drop = TRUE]
       )
-  })
-  d <- purrr::map(d, function(x) {
-    x[-c(1:2), , drop = FALSE]
-  })
+    })
+  d <-
+    purrr::map(d, function(x) {
+      x[-c(1:2), , drop = FALSE]
+    })
 
   # rename columns
   d <-
@@ -59,50 +61,56 @@ elco_irms_import_csv <- function(files) {
     })
 
   # remove unneeded columns
-  d <- purrr::map(d, function(x) {
-    x[, !stringr::str_detect(colnames(x), "unknown")]
-  })
+  d <-
+    purrr::map(d, function(x) {
+      x[, !stringr::str_detect(colnames(x), "unknown")]
+    })
 
   # remove unneeded rows
-  d <- purrr::map(d, function(x) {
-    x[-c(1:4), , drop = FALSE]
-  })
+  d <-
+    purrr::map(d, function(x) {
+      x[-c(1:4), , drop = FALSE]
+    })
 
   # reformat time
-  d <- purrr::map(d, function(x) {
-    x$time <- lubridate::parse_date_time(x$time, orders = "dmy %h:%M")
-    x
-  })
+  d <-
+    purrr::map(d, function(x) {
+      x$time <- lubridate::parse_date_time(x$time, orders = "dmy %h:%M")
+      x
+    })
 
   # reformat units and numeric variables (unfortunately, there's no permil in units::udunits)
-  d <- purrr::map(d, function(x) {
-    dplyr::mutate(x,
-                  "measurement_id" = as.integer(.data$measurement_id),
-                  "sample_mass" = quantities::set_quantities(as.numeric(.data$sample_mass), unit = "mg", errors = 0),
-                  "15N_area" = as.numeric(.data$`15N_area`),
-                  "15N" = as.numeric(.data$`15N`),
-                  "13C_area" = as.numeric(.data$`13C_area`),
-                  "13C" = as.numeric(.data$`13C`),
-                  "18O" = as.numeric(.data$`18O`),
-                  "N" = elco_new_elco(quantities::set_quantities(as.numeric(.data$N)/100, unit = "g/g", errors = 0), el_symbol = "N"),
-                  "C" = elco_new_elco(quantities::set_quantities(as.numeric(.data$C)/100, unit = "g/g", errors = 0), el_symbol = "C")
-    )
-  })
+  d <-
+    purrr::map(d, function(x) {
+      dplyr::mutate(x,
+                    "measurement_id" = as.integer(.data$measurement_id),
+                    "sample_mass" = quantities::set_quantities(as.numeric(.data$sample_mass), unit = "mg_sample", errors = 0),
+                    "15N_area" = as.numeric(.data$`15N_area`),
+                    "15N" = as.numeric(.data$`15N`),
+                    "13C_area" = as.numeric(.data$`13C_area`),
+                    "13C" = as.numeric(.data$`13C`),
+                    "18O" = as.numeric(.data$`18O`),
+                    "N" = quantities::set_quantities(as.numeric(.data$N)/100, unit = "g_N/g_sample", errors = 0),
+                    "C" = quantities::set_quantities(as.numeric(.data$C)/100, unit = "g_C/g_sample", errors = 0)
+      )
+    })
 
   # add file number
-  d <- purrr::map2(d, seq_along(d), function(x, i) {
-    dplyr::bind_cols(
-      tibble::tibble(
-        file_id = rep(i, nrow(x))
-      ),
-      x
-    )
-  })
+  d <-
+    purrr::map2(d, seq_along(d), function(x, i) {
+      dplyr::bind_cols(
+        tibble::tibble(
+          file_id = rep(i, nrow(x))
+        ),
+        x
+      )
+    })
 
   # remove awkward space at the beginning of sample labels
-  d <- purrr::map(d, function(x) {
-    dplyr::mutate(x, "sample_label" = stringr::str_remove(.data$sample_label, pattern = "^ "))
-  })
+  d <-
+    purrr::map(d, function(x) {
+      dplyr::mutate(x, "sample_label" = stringr::str_remove(.data$sample_label, pattern = "^ "))
+    })
 
   # merge
   d <- dplyr::bind_rows(d)